Statistics & Vocab pages rendering.
