+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,6 +1,6 @@
 import NextAuth from 'next-auth';
 import authOptions from '@/utils/authOptions';
 
-export const handler = NextAuth.default(authOptions);
+const handler = NextAuth(authOptions);
 
 export { handler as GET, handler as POST };
